+++ src/iaht.h
@@ -9,5 +9,5 @@ iaht iaht_init(unsigned long long size);
 void iaht_store(iaht L, word index, int x, ...);
 iaht_list iaht_get(iaht L, word index);
 void iaht_reset(iaht L, unsigned long long size);
-void iaht_free(iaht L, int size);
+void iaht_free(iaht L, unsigned long long size);
 #endif
